// src/app/core/model/carDto.ts
export interface CarDto {
  codeCar?: number;
  brandCarId: number;
  brandCarString?: number;
  reference: string;
  price: number;
  modelYear: number;
  color: string;
  horsepower: number;
  numberDoor: number;
  engineDisplacement: number;
  transmission: string;
  fuelType: string;
  numberSeats: number;
  traction: number;
  steering: string;
  category: string;
  imagePath: string;
  stock: number;
}